requirements.txt
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
chardet==3.0.4
discord==1.0.1
discord.py==1.6.0
idna==2.10
isort==5.7.0
multidict==5.1.0
pkg-resources==0.0.0
PTable==0.9.2
pycodestyle==2.6.0
python-dotenv==0.15.0
requests==2.25.1
typing-extensions==3.7.4.3
urllib3==1.26.3
yarl==1.6.3
